add hotfix for (sym)links creating bug

+++ modules.go
@@ -80,7 +80,7 @@ func unTar(r io.Reader, targetBaseDir string) {
 
 		case tar.TypeSymlink:
 			link, _ := os.Readlink(targetFilename)
-			if link != header.Linkname {
+			if link != header.Linkname && fileExists(targetFilename) {
 				if err = os.Remove(targetFilename); err != nil {
 					Fatalf(funcName + "(): error while removing existing file " + targetFilename + " to be replaced with symlink pointing to " + header.Linkname + " Error: " + err.Error())
 				}
@@ -91,7 +91,7 @@ func unTar(r io.Reader, targetBaseDir string) {
 
 		case tar.TypeLink:
 			link, _ := os.Readlink(targetFilename)
-			if link != header.Linkname {
+			if link != header.Linkname && fileExists(targetFilename) {
 				if err = os.Remove(targetFilename); err != nil {
 					Fatalf(funcName + "(): error while removing existing file " + targetFilename + " to be replaced with hardlink pointing to " + header.Linkname + " Error: " + err.Error())
 				}
